requirements.txt
fastapi[standard]==0.115.12
uvicorn[standard]>=0.12.0,<0.14.0
python-dotenv>=1.0.1
pydantic>=2.7.0,<3.0.0
langchain-core==0.3.59
langchain-huggingface==0.2.0
langchain-community==0.3.24
langchain-text-splitters>=0.0.1
langchain-groq>=0.0.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
